Applied patch by Arran - "Patch for a compile warning + getPedTotalAmmo improvement"
mantis: http://bugs.mtasa.com/view.php?id=8646

+++ trunk/MTA10/mods/shared_logic/CClientPed.cpp
@@ -3645,7 +3645,7 @@ void CClientPed::_DestroyModel ()
                 CWeapon * pWeapon = GetWeapon ( m_WeaponTypes [ i ] );
                 if ( pWeapon )
                 {
-                    m_usWeaponAmmo [ i ] = pWeapon->GetAmmoTotal ( );
+                    m_usWeaponAmmo [ i ] = static_cast < ushort > ( pWeapon->GetAmmoTotal ( ) );
                 }
             }
         }
